requirements.txt
altgraph==0.17.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
customtkinter==5.2.2
darkdetect==0.8.0
lxml==5.3.0
macholib==1.16.3
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdfium2==4.30.0
python-docx==1.1.2
typing_extensions==4.12.2
